Small refactorings and tidy-ups.
